feat(user): add GET /users/me endpoint and anime unreviewed filter

+++ src/main/java/com/lucafu/anime_tracker_api/modules/anime/dto/AnimeResponseDto.java
@@ -14,5 +14,5 @@ public class AnimeResponseDto {
     private String imageUrl;
     private Boolean classic;
     private BigDecimal averageScore;
-    private Integer userScore;
+    private BigDecimal userScore;
 }
